+++ README.md
@@ -9,37 +9,36 @@
     $ bundle
 
 ## Использование
-Для сервиса Sms.ru вам необходимо сконфигурировать его
-
-Например:
+Для сервиса Sms.ru вам необходимо сконфигурировать его.
 
     Smsru.configure do |conf|
-      conf.mail = 'blablabla@email.com'
-      conf.api_id = "e7f6f922-838e-c924-79f9-3c4314f63zbd"
-      conf.from = 'foo'
+      conf.mail = 'your-mail'
+      conf.api_id = 'your-api-id'
+      conf.from = 'sender-name'
     end
 
 mail - Ваш Email<br>
-api_id - спец. токен для сервиса<br>
-from - строка отправителя (номер или согласованное название)
+api_id - Спец. токен для сервиса<br>
+from - Строка отправителя (номер или согласованное название)
 
-Метод отправки сообщения
+Отправка сообщения для одного адресата:
 
-    Smsru::Sender.single_sms('79539136846', 'test gem')
+    Smsru::API.send_sms(to, text, api_id, from, test)
 
-Пример отправки сообщения
+to - номер телефона получателя;<br>
+text - текст сообщения в кодировке UTF-8;<br>
+api_id и from - необязательные параметры, если модуль был сконфигурирован ранее;<br>
+test - необязательный параметр, указывающий на то, что сообщение - тестовое (по-умолчанию false)
 
-    Smsru::Sender.single_sms('79539136846', 'test gem')
+Групповая рассылка:
 
-Пример для тестового сообщения
+    Smsru::API.group_send(to, text, api_id, from, test)
 
-    Smsru::Sender.test_sms('79539136846', 'test gem')
+to - массив с номерами телефонов; функция отправляет запросы по 100 штук
 
-Групповая рассылка (пока не протестировано на людях)
+Запрос баланса:
 
-    Smsru::Sender.group_send(to, text, from)
-
-to - массив произвольного размера, функция отправляет запросы по 100 штук; остальные параметры те же
+    Smsru::API.balance(api_id)
 
 ## Contributing
