+++ auto.sh
@@ -3,4 +3,4 @@ git commit -m "1sf-x"
 git push
 sudo gem uninstall smsru
 gem build smsru.gemspec
-sudo gem install smsru-0.0.3.gem
+sudo gem install smsru-0.0.4.gem
