+++ lib/smsru/status.rb
@@ -15,6 +15,7 @@ module Smsru
       "209" =>	"Вы добавили этот номер (или один из номеров) в стоп-лист",
       "210" =>	"Используется GET, где необходимо использовать POST",	
       "211" =>	"Метод не найден",
+      "212" =>  "Текст сообщения необходимо передать в кодировке UTF-8 (вы передали в другой кодировке)",
       "220" =>	"Сервис временно недоступен, попробуйте чуть позже.",
       "230" =>	"Сообщение не принято к отправке, так как на один номер в день нельзя отправлять более 100 сообщений.",
       "300" =>	"Неправильный token (возможно истек срок действия, либо ваш IP изменился)",
